+++ CHANGELOG.md
@@ -25,6 +25,15 @@
     - Add prefix argument to TIFFStackReader to load a subset of TIFF files in
     a folder (#2239)
     - Update ASTRA interface to `direct_FP3D/BP3D` removing copies for GPU `ProjectionOperator` calls (#2134)
+  - Removes the following code which had been deprecated since v24.3.0 or earlier:
+    - Removes `max_iteration` and `log_file` input parameters to `Algorithm`s. 
+    - Removes `max_iteration_stop_criterion`, `objective_to_string`, `verbose_output` and `verbose_header` methods from `Algorithm`s.
+    - Removes `print_interval` and `callback` kwargs from `Algorithm.run` (note: `callbacks` kwarg remains)
+    - Removes `tolerance` input parameter to `CGLS`.
+    - Removes `should_stop` and `flag` methods from `CGLS`.
+    - Removes `alpha`, `beta`, `rtol` and `atol` kwargs from `GD`.
+    - Removes `should_stop` and `objective_function` methods from `GD`.
+    - Removes `norms` and `prob` kwargs from `SPDHG`.
 
 * 25.0.0
   - New features: